add historicalBalance to available metrics
