+++ src/ducks/Signals/chart/preview/NoSignalPreview.module.scss
@@ -1,4 +1,4 @@
-@import "~@santiment-network/ui/mixins";
+@import "../../../../../node_modules/@santiment-network/ui/mixins";
 
 .noDataImg {
   width: 100%;
