+++ src/ducks/Signals/chart/preview/SignalPreview.module.scss
@@ -1,4 +1,4 @@
-@import "~@santiment-network/ui/mixins.scss";
+@import "../../../../../node_modules/@santiment-network/ui/mixins";
 
 .fired {
   @include text('body-3');
