+++ src/ducks/Signals/signalFormManager/sharedForm/SharedTriggerForm.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import Button from '@santiment-network/ui/Button'
 import { couldShowChart } from '../../utils/utils'
-import SignalPreview from '../../chart/SignalPreview'
+import SignalPreview from '../../chart/preview/SignalPreview'
 import SignalCard from '../../../../components/SignalCard/card/SignalCard'
-import NoSignalPreview from '../../chart/NoSignalPreview'
+import NoSignalPreview from '../../chart/preview/NoSignalPreview'
 import { DesktopOnly, MobileOnly } from '../../../../components/Responsive'
 import styles from './ShareTriggerForm.module.scss'
 import CopySignal from '../../../../components/SignalCard/controls/CopySignal'
