+++ src/pages/SonarFeed/ActivityRenderer/ActivityWithBacktesting.js
@@ -4,7 +4,7 @@ import Markdown from 'react-markdown'
 import { Link } from 'react-router-dom'
 import { SignalTypeIcon } from '../../../components/SignalCard/controls/SignalControls'
 import { SIGNAL_ANCHORS } from '../../../ducks/Signals/common/constants'
-import SignalPreview from '../../../ducks/Signals/chart/SignalPreview'
+import SignalPreview from '../../../ducks/Signals/chart/preview/SignalPreview'
 import styles from './ActivityRenderer.module.scss'
 import CopySignal from '../../../components/SignalCard/controls/CopySignal'
 import SignalCreator from '../../../components/SignalCard/card/creator/SignalCreator'
